refactor(orchestrator): extract security inspector workflows
