+++ services/local-service/internal/orchestrator/security.go
@@ -1,0 +1,822 @@
+package orchestrator
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"strings"
+	"time"
+
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/audit"
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/checkpoint"
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/execution"
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/risk"
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/runengine"
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/storage"
+)
+
+// SecurityPendingList handles `agent.security.pending.list` and keeps the
+// pending-authorization list aligned with the merged task-centric read model.
+func (s *Service) SecurityPendingList(params map[string]any) (map[string]any, error) {
+	limit := clampListLimit(intValue(params, "limit", 20))
+	offset := clampListOffset(intValue(params, "offset", 0))
+	unfinishedTasks := newTaskQueryViews(s).tasks("unfinished", "updated_at", "desc")
+	items := pendingApprovalsFromTasks(unfinishedTasks)
+	total := len(items)
+
+	// Keep the legacy runtime response as a safety net when runtime approval
+	// requests exist but the task snapshots do not expose a structured payload.
+	if total == 0 {
+		if s.storage != nil {
+			storedRecords, storedTotal, err := s.storage.ApprovalRequestStore().ListPendingApprovalRequests(context.Background(), limit, offset)
+			if err == nil && storedTotal > 0 {
+				items = approvalRequestRecordsToItems(storedRecords)
+				total = storedTotal
+			} else {
+				runtimeItems, runtimeTotal := s.runEngine.PendingApprovalRequests(limit, offset)
+				items = runtimeItems
+				total = runtimeTotal
+			}
+		} else {
+			runtimeItems, runtimeTotal := s.runEngine.PendingApprovalRequests(limit, offset)
+			items = runtimeItems
+			total = runtimeTotal
+		}
+	} else if offset >= total {
+		items = []map[string]any{}
+	} else {
+		end := offset + limit
+		if end > total {
+			end = total
+		}
+		items = items[offset:end]
+	}
+
+	return map[string]any{
+		"items": items,
+		"page":  pageMap(limit, offset, total),
+	}, nil
+}
+
+// SecurityAuditList returns audit records for security views without mutating
+// recovery or authorization state.
+func (s *Service) SecurityAuditList(params map[string]any) (map[string]any, error) {
+	limit := clampListLimit(intValue(params, "limit", 20))
+	offset := clampListOffset(intValue(params, "offset", 0))
+	taskID := stringValue(params, "task_id", "")
+	if strings.TrimSpace(taskID) == "" {
+		return nil, errors.New("task_id is required")
+	}
+	if s.storage == nil {
+		return map[string]any{"items": []map[string]any{}, "page": pageMap(limit, offset, 0)}, nil
+	}
+	runIDFilter := ""
+	task := runengine.TaskRecord{}
+	if loadedTask, ok := formalReadTask(taskID, s.runEngine, s.taskDetailFromStorage); ok {
+		task = loadedTask
+		runIDFilter = taskAttemptRunIDFilter(task)
+	}
+	records, total, err := s.storage.AuditStore().ListAuditRecords(context.Background(), taskID, runIDFilter, limit, offset)
+	if err != nil {
+		return nil, fmt.Errorf("%w: %v", ErrStorageQueryFailed, err)
+	}
+	if total == 0 && runIDFilter != "" && len(task.AuditRecords) > 0 {
+		items := paginateTaskAuditItems(task.AuditRecords, limit, offset)
+		return map[string]any{
+			"items": items,
+			"page":  pageMap(limit, offset, len(task.AuditRecords)),
+		}, nil
+	}
+	items := make([]map[string]any, 0, len(records))
+	for _, record := range records {
+		items = append(items, record.Map())
+	}
+	return map[string]any{
+		"items": items,
+		"page":  pageMap(limit, offset, total),
+	}, nil
+}
+
+func paginateTaskAuditItems(items []map[string]any, limit, offset int) []map[string]any {
+	if len(items) == 0 || offset >= len(items) {
+		return []map[string]any{}
+	}
+	end := len(items)
+	if limit > 0 && offset+limit < end {
+		end = offset + limit
+	}
+	return cloneMapSlice(items[offset:end])
+}
+
+// SecurityRestorePointsList returns recovery points associated with a task so
+// callers can inspect restore options before requesting an apply action.
+func (s *Service) SecurityRestorePointsList(params map[string]any) (map[string]any, error) {
+	limit := clampListLimit(intValue(params, "limit", 20))
+	offset := clampListOffset(intValue(params, "offset", 0))
+	taskID := stringValue(params, "task_id", "")
+	if s.storage == nil {
+		return map[string]any{"items": []map[string]any{}, "page": pageMap(limit, offset, 0)}, nil
+	}
+	points, total, err := s.storage.RecoveryPointStore().ListRecoveryPoints(context.Background(), taskID, limit, offset)
+	if err != nil {
+		return nil, fmt.Errorf("%w: %v", ErrStorageQueryFailed, err)
+	}
+	items := make([]map[string]any, 0, len(points))
+	for _, point := range points {
+		items = append(items, map[string]any{
+			"recovery_point_id": point.RecoveryPointID,
+			"task_id":           point.TaskID,
+			"summary":           point.Summary,
+			"created_at":        point.CreatedAt,
+			"objects":           append([]string(nil), point.Objects...),
+		})
+	}
+	return map[string]any{
+		"items": items,
+		"page":  pageMap(limit, offset, total),
+	}, nil
+}
+
+// SecurityRestoreApply routes recovery through the same governance path as
+// other risky actions instead of directly mutating workspace state.
+func (s *Service) SecurityRestoreApply(params map[string]any) (map[string]any, error) {
+	recoveryPointID := stringValue(params, "recovery_point_id", "")
+	if strings.TrimSpace(recoveryPointID) == "" {
+		return nil, errors.New("recovery_point_id is required")
+	}
+	taskID := stringValue(params, "task_id", "")
+	point, err := s.findRecoveryPointFromStorage(taskID, recoveryPointID)
+	if err != nil {
+		return nil, err
+	}
+	resolvedTaskID := firstNonEmptyString(strings.TrimSpace(taskID), point.TaskID)
+	task, ok := s.runEngine.GetTask(resolvedTaskID)
+	if !ok {
+		persistedTask, found := s.taskDetailFromStorage(resolvedTaskID)
+		if !found {
+			return nil, ErrTaskNotFound
+		}
+		task = s.runEngine.HydrateTaskFromStorage(persistedTask)
+	}
+
+	recoveryPoint := recoveryPointMap(point)
+	assessment := restoreApplyAssessment(point)
+	pendingExecution := buildRestoreApplyPendingExecution(point, assessment)
+	approvalRequest := buildApprovalRequest(task.TaskID, task.Intent, assessment)
+	bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", "恢复点回滚属于高风险操作，请先确认授权。", time.Now().Format(dateTimeLayout))
+	updatedTask, ok := s.runEngine.MarkWaitingApprovalWithPlan(task.TaskID, approvalRequest, pendingExecution, bubble)
+	if !ok {
+		return nil, ErrTaskNotFound
+	}
+	if err := s.persistApprovalRequestState(updatedTask.TaskID, approvalRequest, assessment.ImpactScope); err != nil {
+		return nil, err
+	}
+	return map[string]any{
+		"applied":        false,
+		"task":           taskMap(updatedTask),
+		"recovery_point": recoveryPoint,
+		"audit_record":   nil,
+		"bubble_message": bubble,
+	}, nil
+}
+
+func (s *Service) applyRestoreAfterApproval(task runengine.TaskRecord, point checkpoint.RecoveryPoint) (runengine.TaskRecord, map[string]any, map[string]any, error) {
+	recoveryPoint := recoveryPointMap(point)
+	applied := false
+	securityStatus := "recovered"
+	finalStatus := "completed"
+	bubbleText := fmt.Sprintf("已根据恢复点 %s 恢复 %d 个对象。", point.RecoveryPointID, len(point.Objects))
+	if s.executor == nil {
+		securityStatus = "execution_error"
+		finalStatus = "failed"
+		bubbleText = "恢复失败：执行后端不可用。"
+	} else if applyResult, err := s.executor.ApplyRecoveryPoint(context.Background(), point); err != nil {
+		securityStatus = "execution_error"
+		finalStatus = "failed"
+		bubbleText = "恢复失败：恢复点内容不可用或恢复执行失败。"
+	} else {
+		applied = true
+		if len(applyResult.RestoredObjects) > 0 {
+			bubbleText = fmt.Sprintf("已根据恢复点 %s 恢复 %d 个对象。", point.RecoveryPointID, len(applyResult.RestoredObjects))
+		}
+	}
+
+	bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", bubbleText, time.Now().Format(dateTimeLayout))
+	updatedTask, ok := s.runEngine.ApplyRecoveryOutcome(task.TaskID, finalStatus, securityStatus, recoveryPoint, bubble)
+	if !ok {
+		return runengine.TaskRecord{}, nil, nil, ErrTaskNotFound
+	}
+	auditRecord := s.writeRestoreAuditRecord(updatedTask.TaskID, updatedTask.RunID, point, applied, bubbleText)
+	updatedTask = s.appendAuditData(updatedTask, compactAuditRecords(auditRecord), nil)
+	return updatedTask, bubble, map[string]any{
+		"applied":        applied,
+		"task":           taskMap(updatedTask),
+		"recovery_point": recoveryPoint,
+		"audit_record":   auditRecord,
+		"bubble_message": bubble,
+	}, nil
+}
+
+func clampListLimit(limit int) int {
+	if limit <= 0 {
+		return 20
+	}
+	if limit > 100 {
+		return 100
+	}
+	return limit
+}
+
+func clampListOffset(offset int) int {
+	if offset < 0 {
+		return 0
+	}
+	return offset
+}
+
+// SecurityRespond handles agent.security.respond. It is the single resume
+// entrypoint for risk-gated tasks, so it must translate allow/deny decisions
+// into runtime state changes, delivery continuation, impact scope reporting,
+// and audit data in one place instead of letting transports or callers stitch
+// those pieces together inconsistently.
+func (s *Service) SecurityRespond(params map[string]any) (map[string]any, error) {
+	taskID := stringValue(params, "task_id", "")
+	task, ok := s.runEngine.GetTask(taskID)
+	if !ok {
+		return nil, ErrTaskNotFound
+	}
+	approvalID, ok := s.activeApprovalIDForTask(task)
+	if !ok {
+		return nil, ErrTaskStatusInvalid
+	}
+
+	decision := stringValue(params, "decision", "allow_once")
+	rememberRule := boolValue(params, "remember_rule", false)
+	authorizationRecord := map[string]any{
+		"authorization_record_id": fmt.Sprintf("auth_%s_%d", task.TaskID, time.Now().UnixNano()),
+		"task_id":                 task.TaskID,
+		"run_id":                  task.RunID,
+		"approval_id":             approvalID,
+		"decision":                decision,
+		"remember_rule":           rememberRule,
+		"operator":                "user",
+		"created_at":              time.Now().Format(dateTimeLayout),
+	}
+	if err := s.persistAuthorizationState(task, authorizationRecord); err != nil {
+		return nil, err
+	}
+	pendingExecution, ok := s.runEngine.PendingExecutionPlan(task.TaskID)
+	if !ok {
+		pendingExecution = s.buildPendingExecution(task, task.Intent)
+	}
+	pendingExecution = s.applyResolvedDeliveryToPlan(task, pendingExecution, task.Intent)
+	impactScope := s.buildImpactScope(task, pendingExecution)
+	operationName := stringValue(pendingExecution, "operation_name", "")
+	if decision == "deny_once" {
+		bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", "已拒绝本次操作，任务已取消。", task.UpdatedAt.Format(dateTimeLayout))
+		updatedTask, ok := s.runEngine.DenyAfterApproval(task.TaskID, authorizationRecord, impactScope, bubble)
+		if !ok {
+			return nil, ErrTaskNotFound
+		}
+		updatedTask = s.appendAuditData(updatedTask, compactAuditRecords(s.audit.BuildAuthorizationAudit(updatedTask.TaskID, updatedTask.RunID, decision, impactScope)), nil)
+		if queueErr := s.drainSessionQueue(updatedTask.SessionID); queueErr != nil {
+			return nil, queueErr
+		}
+		return map[string]any{
+			"authorization_record": authorizationRecord,
+			"task":                 taskMap(updatedTask),
+			"bubble_message":       bubble,
+			"impact_scope":         impactScope,
+		}, nil
+	}
+
+	resumeBubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", "已允许本次操作，任务继续执行。", task.UpdatedAt.Format(dateTimeLayout))
+	processingTask, ok := s.runEngine.ResumeAfterApproval(task.TaskID, authorizationRecord, impactScope, resumeBubble)
+	if !ok {
+		return nil, ErrTaskNotFound
+	}
+	processingTask = s.appendAuditData(processingTask, compactAuditRecords(s.audit.BuildAuthorizationAudit(processingTask.TaskID, processingTask.RunID, decision, impactScope)), nil)
+	if operationName == "restore_apply" {
+		recoveryPointID := stringValue(pendingExecution, "recovery_point_id", "")
+		point, err := s.findRecoveryPointFromStorage(task.TaskID, recoveryPointID)
+		if err != nil {
+			return nil, err
+		}
+		updatedTask, _, response, err := s.applyRestoreAfterApproval(processingTask, point)
+		if err != nil {
+			return nil, err
+		}
+		return map[string]any{
+			"authorization_record": authorizationRecord,
+			"task":                 taskMap(updatedTask),
+			"bubble_message":       response["bubble_message"],
+			"impact_scope":         impactScope,
+			"delivery_result":      nil,
+			"recovery_point":       response["recovery_point"],
+			"audit_record":         response["audit_record"],
+			"applied":              response["applied"],
+		}, nil
+	}
+	if stringValue(pendingExecution, "kind", "") == "screen_analysis" {
+		updatedTask, bubble, deliveryResult, err := s.executeScreenAnalysisAfterApproval(processingTask, pendingExecution)
+		if err != nil {
+			return nil, err
+		}
+		if updatedTask.Status == "completed" {
+			updatedTask, _ = s.runEngine.ResolveAuthorization(task.TaskID, authorizationRecord, impactScope)
+		}
+		if taskIsTerminal(updatedTask.Status) {
+			if queueErr := s.drainSessionQueue(updatedTask.SessionID); queueErr != nil {
+				return nil, queueErr
+			}
+		}
+		return map[string]any{
+			"authorization_record": authorizationRecord,
+			"task":                 taskMap(updatedTask),
+			"bubble_message":       bubble,
+			"impact_scope":         impactScope,
+			"delivery_result":      deliveryResult,
+		}, nil
+	}
+
+	updatedTask, resultBubble, deliveryResult, _, err := s.executeTask(processingTask, snapshotFromTask(processingTask), processingTask.Intent)
+	if err != nil {
+		return nil, err
+	}
+	if updatedTask.Status == "completed" {
+		updatedTask, _ = s.runEngine.ResolveAuthorization(task.TaskID, authorizationRecord, impactScope)
+	}
+	if updatedTask.Status == "failed" {
+		deliveryResult = nil
+	}
+	if taskIsTerminal(updatedTask.Status) {
+		if queueErr := s.drainSessionQueue(updatedTask.SessionID); queueErr != nil {
+			return nil, queueErr
+		}
+	}
+
+	response := map[string]any{
+		"authorization_record": authorizationRecord,
+		"task":                 taskMap(updatedTask),
+		"bubble_message":       resultBubble,
+		"impact_scope":         impactScope,
+	}
+	if len(deliveryResult) > 0 {
+		response["delivery_result"] = deliveryResult
+	} else {
+		response["delivery_result"] = nil
+	}
+	return response, nil
+}
+
+func aggregateRiskLevel(tasks []runengine.TaskRecord, pendingApprovals []map[string]any, fallback string) string {
+	if len(pendingApprovals) > 0 {
+		return "red"
+	}
+	result := fallback
+	for _, task := range tasks {
+		switch task.RiskLevel {
+		case "red":
+			return "red"
+		case "yellow":
+			result = "yellow"
+		case "green":
+			if result == "" {
+				result = "green"
+			}
+		}
+	}
+	if result == "" {
+		return "green"
+	}
+	return result
+}
+
+func aggregateSecurityStatus(tasks []runengine.TaskRecord, pendingTotal int) string {
+	if pendingTotal > 0 {
+		return "pending_confirmation"
+	}
+	for _, task := range tasks {
+		status := stringValue(task.SecuritySummary, "security_status", "")
+		if status != "" && status != "normal" {
+			return status
+		}
+	}
+	return "normal"
+}
+
+func latestAuditRecordFromTasks(tasks []runengine.TaskRecord) map[string]any {
+	var latestAudit map[string]any
+	var latestAt time.Time
+	for _, task := range tasks {
+		for _, auditRecord := range task.AuditRecords {
+			auditAt := parseAuditTime(auditRecord)
+			if latestAudit == nil || auditAt.After(latestAt) {
+				latestAudit = cloneMap(auditRecord)
+				latestAt = auditAt
+			}
+		}
+	}
+	return latestAudit
+}
+
+func (s *Service) latestAuditRecordFromStorage(taskID string) map[string]any {
+	if s.storage == nil {
+		return nil
+	}
+	items, _, err := s.storage.AuditStore().ListAuditRecords(context.Background(), taskID, "", 1, 0)
+	if err != nil || len(items) == 0 {
+		return nil
+	}
+	return normalizeTaskDetailAuditRecord(taskID, items[0].Map())
+}
+
+func (s *Service) loadAttemptAuditRecordsFromStorage(task runengine.TaskRecord, limit, offset int) []map[string]any {
+	if s == nil || s.storage == nil || s.storage.AuditStore() == nil || strings.TrimSpace(task.TaskID) == "" {
+		return nil
+	}
+	items, _, err := s.storage.AuditStore().ListAuditRecords(context.Background(), task.TaskID, taskAttemptRunIDFilter(task), limit, offset)
+	if err != nil {
+		return nil
+	}
+	result := make([]map[string]any, 0, len(items))
+	for _, item := range items {
+		result = append(result, item.Map())
+	}
+	return result
+}
+
+func (s *Service) latestToolCallFromStorage(taskID, runID string) map[string]any {
+	if s == nil || s.storage == nil || s.storage.ToolCallSink() == nil || strings.TrimSpace(taskID) == "" {
+		return nil
+	}
+	items, _, err := s.storage.ToolCallStore().ListToolCalls(context.Background(), taskID, runID, 1, 0)
+	if err != nil || len(items) == 0 {
+		return nil
+	}
+	item := items[0]
+	return map[string]any{
+		"tool_call_id": item.ToolCallID,
+		"run_id":       item.RunID,
+		"task_id":      item.TaskID,
+		"step_id":      item.StepID,
+		"tool_name":    item.ToolName,
+		"status":       item.Status,
+		"input":        cloneMap(item.Input),
+		"output":       cloneMap(item.Output),
+		"error_code":   item.ErrorCode,
+		"duration_ms":  item.DurationMS,
+	}
+}
+
+func parseAuditTime(auditRecord map[string]any) time.Time {
+	createdAt := stringValue(auditRecord, "created_at", "")
+	if createdAt == "" {
+		return time.Time{}
+	}
+	parsed, err := time.Parse(time.RFC3339Nano, createdAt)
+	if err != nil {
+		return time.Time{}
+	}
+	return parsed
+}
+
+func latestRestorePointFromTasks(tasks []runengine.TaskRecord) map[string]any {
+	for _, task := range tasks {
+		restorePoint, ok := task.SecuritySummary["latest_restore_point"].(map[string]any)
+		if ok && len(restorePoint) > 0 {
+			return cloneMap(restorePoint)
+		}
+	}
+	return nil
+}
+
+func latestRestorePointFromSummary(summary map[string]any) map[string]any {
+	if summary == nil {
+		return nil
+	}
+	latestRestorePoint, ok := summary["latest_restore_point"].(map[string]any)
+	if !ok {
+		return nil
+	}
+	return cloneMap(latestRestorePoint)
+}
+
+func activeTaskDetailApprovalRequest(task runengine.TaskRecord) map[string]any {
+	if task.Status != "waiting_auth" || len(task.ApprovalRequest) == 0 {
+		return nil
+	}
+	return normalizeTaskDetailApprovalRequest(task.TaskID, task.RiskLevel, task.ApprovalRequest)
+}
+
+func (s *Service) normalizeTaskDetailRestorePoint(taskID string, securitySummary map[string]any) map[string]any {
+	if latestRestorePoint := normalizeTaskDetailRecoveryPoint(taskID, latestRestorePointFromSummary(securitySummary)); latestRestorePoint != nil {
+		return latestRestorePoint
+	}
+	if restorePoint := s.latestRestorePointFromStorage(taskID); restorePoint != nil {
+		return restorePoint
+	}
+	return nil
+}
+
+func normalizeTaskDetailApprovalRequest(taskID, fallbackRiskLevel string, approvalRequest map[string]any) map[string]any {
+	if len(approvalRequest) == 0 {
+		return nil
+	}
+
+	approvalID := strings.TrimSpace(stringValue(approvalRequest, "approval_id", ""))
+	approvalTaskID := strings.TrimSpace(stringValue(approvalRequest, "task_id", ""))
+	operationName := strings.TrimSpace(stringValue(approvalRequest, "operation_name", ""))
+	targetObject := strings.TrimSpace(stringValue(approvalRequest, "target_object", ""))
+	reason := strings.TrimSpace(stringValue(approvalRequest, "reason", ""))
+	status := strings.TrimSpace(stringValue(approvalRequest, "status", ""))
+	createdAt := strings.TrimSpace(stringValue(approvalRequest, "created_at", ""))
+	riskLevel := strings.TrimSpace(stringValue(approvalRequest, "risk_level", ""))
+	if riskLevel == "" {
+		riskLevel = strings.TrimSpace(fallbackRiskLevel)
+	}
+
+	if approvalID == "" || approvalTaskID != taskID || operationName == "" || targetObject == "" || reason == "" || createdAt == "" {
+		return nil
+	}
+	if status != "pending" || !isSupportedRiskLevel(riskLevel) {
+		return nil
+	}
+
+	return map[string]any{
+		"approval_id":    approvalID,
+		"task_id":        approvalTaskID,
+		"operation_name": operationName,
+		"risk_level":     riskLevel,
+		"target_object":  targetObject,
+		"reason":         reason,
+		"status":         status,
+		"created_at":     createdAt,
+	}
+}
+
+func normalizeTaskDetailRecoveryPoint(taskID string, recoveryPoint map[string]any) map[string]any {
+	if len(recoveryPoint) == 0 {
+		return nil
+	}
+
+	recoveryPointID := strings.TrimSpace(stringValue(recoveryPoint, "recovery_point_id", ""))
+	recoveryTaskID := strings.TrimSpace(stringValue(recoveryPoint, "task_id", ""))
+	summary := strings.TrimSpace(stringValue(recoveryPoint, "summary", ""))
+	createdAt := strings.TrimSpace(stringValue(recoveryPoint, "created_at", ""))
+	objects, ok := normalizeStringSlice(recoveryPoint["objects"])
+	if !ok {
+		return nil
+	}
+
+	if recoveryPointID == "" || recoveryTaskID != taskID || summary == "" || createdAt == "" {
+		return nil
+	}
+
+	return map[string]any{
+		"recovery_point_id": recoveryPointID,
+		"task_id":           recoveryTaskID,
+		"summary":           summary,
+		"created_at":        createdAt,
+		"objects":           objects,
+	}
+}
+
+func isSupportedRiskLevel(riskLevel string) bool {
+	switch riskLevel {
+	case "green", "yellow", "red":
+		return true
+	default:
+		return false
+	}
+}
+
+func normalizeStringSlice(value any) ([]string, bool) {
+	switch typed := value.(type) {
+	case []string:
+		return append([]string(nil), typed...), true
+	case []any:
+		items := make([]string, 0, len(typed))
+		for _, item := range typed {
+			text, ok := item.(string)
+			if !ok {
+				return nil, false
+			}
+			items = append(items, text)
+		}
+		return items, true
+	default:
+		return nil, false
+	}
+}
+
+func (s *Service) latestRestorePointFromStorage(taskID string) map[string]any {
+	if s.storage == nil {
+		return nil
+	}
+	items, _, err := s.storage.RecoveryPointStore().ListRecoveryPoints(context.Background(), taskID, 1, 0)
+	if err != nil || len(items) == 0 {
+		return nil
+	}
+	item := items[0]
+	return map[string]any{
+		"recovery_point_id": item.RecoveryPointID,
+		"task_id":           item.TaskID,
+		"summary":           item.Summary,
+		"created_at":        item.CreatedAt,
+		"objects":           append([]string(nil), item.Objects...),
+	}
+}
+
+func (s *Service) findRecoveryPointFromStorage(taskID, recoveryPointID string) (checkpoint.RecoveryPoint, error) {
+	if s.storage == nil {
+		return checkpoint.RecoveryPoint{}, fmt.Errorf("%w: recovery point store unavailable", ErrStorageQueryFailed)
+	}
+	item, err := s.storage.RecoveryPointStore().GetRecoveryPoint(context.Background(), recoveryPointID)
+	if err != nil {
+		if errors.Is(err, storage.ErrRecoveryPointNotFound) {
+			return checkpoint.RecoveryPoint{}, ErrRecoveryPointNotFound
+		}
+		return checkpoint.RecoveryPoint{}, fmt.Errorf("%w: %v", ErrStorageQueryFailed, err)
+	}
+	if taskID != "" && item.TaskID != taskID {
+		return checkpoint.RecoveryPoint{}, ErrRecoveryPointNotFound
+	}
+	return item, nil
+}
+
+func recoveryPointMap(point checkpoint.RecoveryPoint) map[string]any {
+	return map[string]any{
+		"recovery_point_id": point.RecoveryPointID,
+		"task_id":           point.TaskID,
+		"summary":           point.Summary,
+		"created_at":        point.CreatedAt,
+		"objects":           append([]string(nil), point.Objects...),
+	}
+}
+
+func restoreApplyAssessment(point checkpoint.RecoveryPoint) execution.GovernanceAssessment {
+	impactScope := restoreImpactScope(point)
+	return execution.GovernanceAssessment{
+		OperationName:      "restore_apply",
+		TargetObject:       firstNonEmptyString(firstImpactFile(impactScope), firstNonEmptyString(strings.Join(point.Objects, ", "), "workspace")),
+		RiskLevel:          "red",
+		ApprovalRequired:   true,
+		CheckpointRequired: false,
+		Reason:             "policy_requires_authorization",
+		ImpactScope:        impactScope,
+	}
+}
+
+func buildRestoreApplyPendingExecution(point checkpoint.RecoveryPoint, assessment execution.GovernanceAssessment) map[string]any {
+	return map[string]any{
+		"operation_name":      assessment.OperationName,
+		"target_object":       assessment.TargetObject,
+		"risk_level":          assessment.RiskLevel,
+		"risk_reason":         assessment.Reason,
+		"impact_scope":        cloneMap(assessment.ImpactScope),
+		"recovery_point_id":   point.RecoveryPointID,
+		"checkpoint_required": assessment.CheckpointRequired,
+	}
+}
+
+func restoreImpactScope(point checkpoint.RecoveryPoint) map[string]any {
+	files := append([]string(nil), point.Objects...)
+	outOfWorkspace := false
+	for _, filePath := range files {
+		normalized := strings.TrimSpace(strings.ReplaceAll(filePath, "\\", "/"))
+		if normalized == "" {
+			continue
+		}
+		if !strings.HasPrefix(normalized, "workspace/") && normalized != "workspace" {
+			outOfWorkspace = true
+			break
+		}
+	}
+	return map[string]any{
+		"files":                    files,
+		"webpages":                 []string{},
+		"apps":                     []string{},
+		"out_of_workspace":         outOfWorkspace,
+		"overwrite_or_delete_risk": true,
+	}
+}
+
+func firstImpactFile(impactScope map[string]any) string {
+	if len(impactScope) == 0 {
+		return ""
+	}
+	files, ok := impactScope["files"].([]string)
+	if !ok || len(files) == 0 {
+		return ""
+	}
+	return files[0]
+}
+
+func (s *Service) writeRestoreAuditRecord(taskID, runID string, point checkpoint.RecoveryPoint, applied bool, summary string) map[string]any {
+	if s.audit == nil {
+		return nil
+	}
+	input := audit.RecordInput{
+		TaskID:  taskID,
+		RunID:   runID,
+		Type:    "recovery",
+		Action:  "restore_apply",
+		Summary: firstNonEmptyString(strings.TrimSpace(summary), "restore apply completed"),
+		Target:  firstNonEmptyString(strings.Join(point.Objects, ", "), "recovery_scope"),
+		Result:  map[bool]string{true: "success", false: "failed"}[applied],
+	}
+	if record, err := s.audit.Write(context.Background(), input); err == nil {
+		return record.Map()
+	}
+	if record, err := s.audit.BuildRecord(input); err == nil {
+		return record.Map()
+	}
+	return nil
+}
+
+func normalizeTaskDetailAuthorizationRecord(taskID string, authorizationRecord map[string]any) map[string]any {
+	if len(authorizationRecord) == 0 {
+		return nil
+	}
+
+	recordID := strings.TrimSpace(stringValue(authorizationRecord, "authorization_record_id", ""))
+	recordTaskID := strings.TrimSpace(stringValue(authorizationRecord, "task_id", ""))
+	approvalID := strings.TrimSpace(stringValue(authorizationRecord, "approval_id", ""))
+	decision := normalizeTaskDetailAuthorizationDecision(stringValue(authorizationRecord, "decision", ""))
+	operator := strings.TrimSpace(stringValue(authorizationRecord, "operator", ""))
+	createdAt := strings.TrimSpace(stringValue(authorizationRecord, "created_at", ""))
+	if recordID == "" || recordTaskID != taskID || approvalID == "" || decision == "" || operator == "" || createdAt == "" {
+		return nil
+	}
+
+	return map[string]any{
+		"authorization_record_id": recordID,
+		"task_id":                 recordTaskID,
+		"approval_id":             approvalID,
+		"decision":                decision,
+		"remember_rule":           boolValue(authorizationRecord, "remember_rule", false),
+		"operator":                operator,
+		"created_at":              createdAt,
+	}
+}
+
+func normalizeTaskDetailAuthorizationDecision(decision string) string {
+	switch strings.TrimSpace(decision) {
+	case "allow_once", "allow_always":
+		return "allow_once"
+	case "deny_once", "deny_always":
+		return "deny_once"
+	default:
+		return ""
+	}
+}
+
+func normalizeTaskDetailAuditRecord(taskID string, auditRecord map[string]any) map[string]any {
+	if len(auditRecord) == 0 {
+		return nil
+	}
+
+	recordID := strings.TrimSpace(firstNonEmptyString(stringValue(auditRecord, "audit_id", ""), stringValue(auditRecord, "audit_record_id", "")))
+	recordTaskID := strings.TrimSpace(stringValue(auditRecord, "task_id", ""))
+	recordType := strings.TrimSpace(firstNonEmptyString(stringValue(auditRecord, "type", ""), stringValue(auditRecord, "category", "")))
+	action := strings.TrimSpace(stringValue(auditRecord, "action", ""))
+	summary := strings.TrimSpace(firstNonEmptyString(stringValue(auditRecord, "summary", ""), stringValue(auditRecord, "reason", "")))
+	target := strings.TrimSpace(firstNonEmptyString(stringValue(auditRecord, "target", ""), impactScopeTarget(mapValue(auditRecord, "impact_scope"), "")))
+	result := strings.TrimSpace(stringValue(auditRecord, "result", ""))
+	createdAt := strings.TrimSpace(stringValue(auditRecord, "created_at", ""))
+	if recordID == "" || recordTaskID != taskID || recordType == "" || action == "" || summary == "" || target == "" || result == "" || createdAt == "" {
+		return nil
+	}
+
+	return map[string]any{
+		"audit_id":   recordID,
+		"task_id":    recordTaskID,
+		"type":       recordType,
+		"action":     action,
+		"summary":    summary,
+		"target":     target,
+		"result":     result,
+		"created_at": createdAt,
+	}
+}
+
+func governanceInterceptionBubble(assessment execution.GovernanceAssessment) string {
+	switch assessment.Reason {
+	case risk.ReasonOutOfWorkspace:
+		return "目标超出工作区边界，已阻止本次操作。"
+	case risk.ReasonCommandNotAllowed:
+		return "命令存在高危风险，已被策略拦截。"
+	case risk.ReasonCapabilityDenied:
+		return "当前平台能力不可用，已阻止本次操作。"
+	default:
+		return "高风险操作已被策略拦截，未进入执行。"
+	}
+}
+
+func impactScopeTarget(impactScope map[string]any, fallback string) string {
+	if files := stringSliceValue(impactScope["files"]); len(files) > 0 {
+		return files[0]
+	}
+	return firstNonEmptyString(strings.TrimSpace(fallback), "main_flow")
+}
